fix(KEN-815): a program is handed what the reading can name, and the rest scores
Which operand of an interpreter holds a command line is a position on the
line, not a property of the command. The old reading asked whether a
standalone "-c" was anywhere in the words, which missed a launcher standing
in front of the interpreter and an option letter travelling in a bundle,
and read every operand after the command string as code when only the first
is one. The tokenizer now resolves the launcher, reads the bundle the way
the shell reads it, and returns the position.

The other half is which files the reading applies to at all. Every
non-binary file in a skill tree was read as a command line from its first
byte, so a switch a Python or a Rust file passes to a program sat inside
that language's quotes, nothing parsed as argv, and the file scored
nothing. A file is judged shell by its shebang first, then a shell suffix,
then a name carrying no suffix. Everything else is a language this rule
does not read, and every position in it counts rather than passing in
silence.

The whole-tree measurement over the 19 shipped skills is unchanged:
growth-guards 83 on its three test lines, orch 100, review-gate 85,
iced-rs 97, the rest 100.

crates/core/tests/quality/rules.rs was one line under its cap, so the
safety-bypass tests move to rules_bypass.rs.
