+++ changelog.d/fixed/KEN-815.md
@@ -1,3 +1,3 @@
 - The safety score counts a dangerous switch where a line would run it: a
-  README, a hook's message and a `case` arm naming one no longer score, and a
-  switch quoted on a command line still does.
+  README or `case` arm naming one no longer scores, and one on a command
+  line, quoted or in a Python or Rust file, still does.
